% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_funcs.R
\name{all_aggregations}
\alias{all_aggregations}
\title{Applies an aggregation function such as union to a tgraph for a range of
possible aggregation windows}
\usage{
all_aggregations(tgraph, agg_func, max_stepsize)
}
\arguments{
\item{tgraph}{a dynamic network}

\item{agg_func}{a function that transforms two or more igraphs into one}

\item{max_stepsize}{the maximum aggregation step size}
}
\value{
A list of max_stepsize dynamic networks
}
\description{
This function applies apply_aggregator to a range of step_size values
}

