% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_funcs.R
\name{apply_aggregator}
\alias{apply_aggregator}
\title{Applies an aggregation function such as union to a tgraph with a fixed
aggregation window}
\usage{
apply_aggregator(tgraph, agg_func = union, step_size = 2)
}
\arguments{
\item{tgraph}{a dynamic network}

\item{agg_func}{a function that transforms two or more igraphs into one}

\item{step_size}{the aggregation step size}
}
\value{
the dynamic network that is the described aggregation of tgraph
}
\description{
This function divides tgraph into subgroups of length step_size, and applies
agg_func to each
}

