% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_metrics.R
\name{avg_degree}
\alias{avg_degree}
\title{the average degree of the graph}
\usage{
avg_degree(G)
}
\arguments{
\item{G}{an igraph graph object}
}
\value{
the average degree of G
}
\description{
the average degree of the graph
}
