% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{tgraph_n_vertices}
\alias{tgraph_n_vertices}
\title{calculates the number of vertices in a dynamic network}
\usage{
tgraph_n_vertices(tgraph)
}
\arguments{
\item{tgraph}{a dynamic network}
}
\value{
the number of vertices in tgraph
}
\description{
calculates the number of vertices in a dynamic network
}
\examples{
dn <- uniform_rand_tgraph(t=2, n_vertices = 6)
six <- tgraph_n_vertices(dn)
}

