% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{tgraph_plot_metric}
\alias{tgraph_plot_metric}
\title{Plots a given metric at each time step in a dynamic network}
\usage{
tgraph_plot_metric(tgraph, graph_metric = avg_degree)
}
\arguments{
\item{tgraph}{a dynamic network}

\item{graph_metric}{a function that takes an igraph graph and returns a
numeric value}
}
\value{
a plot
}
\description{
Plots a given metric at each time step in a dynamic network
}

