% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgraph_generators.R
\name{edgeless_tgraph}
\alias{edgeless_tgraph}
\title{A tgraph with no edges}
\usage{
edgeless_tgraph(t, n_vertices)
}
\arguments{
\item{t}{the number of time steps}

\item{n_vertices}{the number of vertices}
}
\value{
A tgraph with no edges, t time steps and n_vertices vertices
}
\description{
A tgraph with no edges
}
\examples{
lonely_guy_ten_times <- edgeless_tgraph(10, 1)
five_isolated_people_six_times <- edgeless_tgraph(6, 5)
}
