% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeordered_compatibility.R
\name{get_tgraph_vertex_names}
\alias{get_tgraph_vertex_names}
\title{Returns a string vector of a tgraph's vertices' names. If the vertices
are unnamed, returns the string vector of their indices.}
\usage{
get_tgraph_vertex_names(tgraph)
}
\arguments{
\item{tgraph}{a dynamic network}
}
\value{
A vector of length(dn) strings
}
\description{
Returns a string vector of a tgraph's vertices' names. If the vertices
are unnamed, returns the string vector of their indices.
}
\examples{
dn <- uniform_rand_tgraph(t=12, n_vertices = 6, p_edge = 0.4)
names <- get_tgraph_vertex_names(dn)
# [1] "1" "2" "3" "4" "5" "6"
}

