% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtm.R
\name{group_members}
\alias{group_members}
\title{Group Members}
\usage{
group_members(g, group_n_offset = 0)
}
\arguments{
\item{g}{an igraph graph}

\item{group_n_offset}{optional arg, an offset added to each
group number. Useful in group_time_members.}
}
\value{
a "GM" list---a list of group-member vectors. Each
vector's first element is the group number, and the remaining
elements are the indices of the vertices that are in that group.
Note that for unlabeled igraph vertices the vertex index is
identical to the vertex.
}
\description{
Generates the "GM" info (GTM without the T) for a single graph
}
\examples{
dn <- uniform_rand_tgraph(t=12, n_vertices = 6, p_edge = 0.4)
}
