% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtm.R
\name{group_members_only}
\alias{group_members_only}
\title{Group Members Only}
\usage{
group_members_only(g, group_n_offset = 0)
}
\arguments{
\item{g}{an igraph graph}
}
\value{
a list of vectors, where the ith vector indexes the vertices
in the ith group (connected subgraph) of g
}
\description{
Like group_members, but without explicitly naming the group numbers
(this info instead encoded in the index of each vertex list)
}
\examples{
dn <- uniform_rand_tgraph(t=12, n_vertices = 6, p_edge = 0.4)
}
