% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtm.R
\name{group_time_members}
\alias{group_time_members}
\title{Group Time Members}
\usage{
group_time_members(tgraph)
}
\arguments{
\item{tgraph}{a tgraph}
}
\value{
a list of vectors that is the group-time-member object
describing tgraph. Each vector's first element identifies a group
number in the tgraph. The second element identifies the time when
the group occurs. The rest of the vector indexes the vertices
present in that group.
}
\description{
Generates the group-time-member info for a tgraph, useful in such
community analysis algorithms as CommDy. Nests results by time.
}
\examples{
dn <- uniform_rand_tgraph(t=12, n_vertices = 6, p_edge = 0.4)
}
