% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtm.R
\name{group_time_members_time_nested}
\alias{group_time_members_time_nested}
\title{Group Time Members (nested)}
\usage{
group_time_members_time_nested(tgraph)
}
\arguments{
\item{tgraph}{a tgraph}
}
\value{
a list of lists (of lists). Top level list is time, bottom
level is a line of gtm info
}
\description{
Generates the group-time-member info for a tgraph, useful in such
community analysis algorithms as CommDy. Nests results by time.
}
\examples{
dn <- uniform_rand_tgraph(t=12, n_vertices = 6, p_edge = 0.4)
}
