% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{half_as_close_to_one}
\alias{half_as_close_to_one}
\title{returns the number that is half as close to one as the input number}
\usage{
half_as_close_to_one(x)
}
\arguments{
\item{x}{a number}
}
\value{
1 - (1-x)/2
}
\description{
returns the number that is half as close to one as the input number
}
\examples{
zero_point_seven_five <- half_as_close_to_one(0.5)
two <- half_as_close_to_one(3)
}

