% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgraph_utils.R
\name{has_constant_Vs}
\alias{has_constant_Vs}
\title{Has constant Vertices}
\usage{
has_constant_Vs(tgraph)
}
\arguments{
\item{tgraph}{a dynamic network}
}
\value{
whether all graphs in the tgraph have the same vertices.
}
\description{
Answers whether the graphs in a tgraph share a constant vertex set
}
\examples{
dn <- constant_rand_tgraph(t=12, n_vertices = 6, p_edge = 0.4)
time_ordered <- make_timeordered(dn)
}

