% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtm_io.R
\name{load_gtm}
\alias{load_gtm}
\title{Load GTM}
\usage{
load_gtm(filename)
}
\arguments{
\item{filename}{the name of the .gtm file to be loaded.}
}
\value{
a list of gtm-line vectors
}
\description{
Loads a .gtm file into a list of GTM vectors
}
