% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgraph_utils.R
\name{make_constant_Vs}
\alias{make_constant_Vs}
\title{Make Constant Vertices}
\usage{
make_constant_Vs(tgraph)
}
\arguments{
\item{tgraph}{a dynamic network}
}
\value{
the same tgraph with constant vertices
}
\description{
Makes the graphs in a tgraph share a constant vertex set
}
\examples{
dn <- constant_rand_tgraph(t=12, n_vertices = 6, p_edge = 0.4)
make_constant_Vs(dn)
}
