% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgraph_utils.R
\name{make_named_tgraph}
\alias{make_named_tgraph}
\title{Make Named tgraph}
\usage{
make_named_tgraph(tgraph)
}
\arguments{
\item{tgraph}{a dynamic network}
}
\value{
a tgraph equal to the input tgraph but with all vertices named
}
\description{
If the igraphs making up a tgraph have unnamed vertices, applies trivial
names based on vertex index
}
\examples{
dn <- constant_rand_tgraph(t=12, n_vertices = 6, p_edge = 0.4)
make_named_tgraph(dn)
}
