% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeordered_compatibility.R
\name{make_tgraph}
\alias{make_tgraph}
\title{Converts a timeordered network into a dynamic network}
\usage{
make_tgraph(tonetwork)
}
\arguments{
\item{tgraph}{a dynamic network}
}
\value{
An igraph graph representing tgraph as a time ordered network
}
\description{
Uses the timeordered.generatenetworkslices function with a time window of 1.
}
\examples{
dn <- uniform_rand_tgraph(t=12, n_vertices = 6, p_edge = 0.4)
time_ordered <- make_timeordered(dn)
}
