% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeordered_compatibility.R
\name{make_timeordered}
\alias{make_timeordered}
\title{Converts a dynamic network into a timeordered network
according the conventions of package timeordered.}
\usage{
make_timeordered(tgraph)
}
\arguments{
\item{tgraph}{a dynamic network}
}
\value{
An igraph graph representing tgraph as a time ordered network
}
\description{
Converts a dynamic network into a timeordered network
according the conventions of package timeordered.
}
\examples{
dn <- uniform_rand_tgraph(t=12, n_vertices = 6, p_edge = 0.4)
time_ordered <- make_timeordered(dn)
}
