% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twin.R
\name{metric_vs_agg_step}
\alias{metric_vs_agg_step}
\title{Apply a tgraph metric on a given tgraph and its time-step-wise aggregations}
\usage{
metric_vs_agg_step(tgraph, tgraph_metric, agg_func, max_agg_step)
}
\arguments{
\item{tgraph}{a dynamic network}

\item{tgraph_metric}{a function mapping a dynamic network to a numeric value.}

\item{agg_func}{an aggregation function that takes two or more igraph graphs
and returns a single graph, e.g. union.}

\item{max_step_size}{the largest number of time steps to be aggregated at
once by the aggregation function.}
}
\value{
A max_agg_step - length vector of tgraph_metric applied to the
 aggregations of tgraph.
}
\description{
Apply a tgraph metric on a given tgraph and its time-step-wise aggregations
}
\examples{
dn <- uniform_rand_tgraph(t=12, n_vertices = 6, p_edge = 0.4)
avg_degree_variance_vs_agg_step <- metric_vs_agg_step(
   tgraph = dn,
   tgraph_metric = var_avg_degree
)
}
