% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgraph_generators.R
\name{oversample_distribute}
\alias{oversample_distribute}
\title{Replaces each graph in a tgraph with w graphs, with the orignal graph's edges
randomly assigned.}
\usage{
oversample_distribute(tgraph, w)
}
\arguments{
\item{tgraph}{a dynamic network}

\item{w}{the number of times each time step should be duplicated}
}
\value{
A dynamic network of length w * length(tgraph)
}
\description{
Replaces each graph in a tgraph with w graphs, with the orignal graph's edges
randomly assigned.
}
\examples{
dn <- uniform_rand_tgraph(t=12, n_vertices = 6, p_edge = 0.4)
same_edges_twice_as_long <- oversample_distribute(dn, 2)
}
