% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{rectangularize}
\alias{rectangularize}
\title{Calculate the dimensions of a rectangle with n items and n columns.}
\usage{
rectangularize(n_items, n_cols = 2)
}
\arguments{
\item{n_items}{the integer number of items to be placed in a rectangular grid}

\item{n_cols}{the number of columns (or desired width) of the rectangle}
}
\value{
a vector of length 2 containing the (width, height) of the rectangle
}
\description{
This function tells you the number of rows and columns necessary to arrange
n items in a rectangle.
}
\details{
The bottom row of the rectangle might have fewer items than above rows, but
never zero items.
}
\examples{
rectangularize(3, 2) # c(2,2), a 3x2 grid could hold 3 objects in two columns
rectangularize(7, 2) # c(4,2), a 4x2 grid would hold 7 objects in two columns
rectangularize(8, 2) # c(4,2)
rectangularize(9, 3) # c(3,3)
}
