% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgraph_generators.R
\name{uniform_rand_tgraph}
\alias{uniform_rand_tgraph}
\title{Creates a uniform dynamic network with fixed edge probability.}
\usage{
uniform_rand_tgraph(t, n_vertices, p_edge)
}
\arguments{
\item{t}{the number of time steps, or length, of the dynamic network}

\item{n_vertices}{the number of vertices in the network}

\item{p_edge}{the uniform edge probability between two vertices at any time}
}
\value{
A dynamic network of length t
}
\description{
This function simply creates t random graphs with a fixed, uniform edge
probability.
}
\examples{
dn <- uniform_rand_tgraph(t=12, n_vertices = 6, p_edge = 0.4)
}

