% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgraph_utils.R
\name{vertex_superset}
\alias{vertex_superset}
\title{Vertex Superset}
\usage{
vertex_superset(tgraph)
}
\arguments{
\item{tgraph}{a dynamic network}
}
\value{
a set of igraph vertices like the return of the igraph V function
}
\description{
Returns the set of all igraph vertices present at any time in the tgraph
}
\examples{
dn <- constant_rand_tgraph(t=12, n_vertices = 6, p_edge = 0.4)
make_constant_Vs(dn)
}
